Upgrade to .NET 10 on Iteration 2

+++ net472/eShopLegacyMVC/Controllers/CatalogController.cs
@@ -1,44 +1,43 @@
-﻿using System.Collections.Generic;
-using System.Net;
-using System.Web.Mvc;
+using System.Collections.Generic;
+using Microsoft.AspNetCore.Mvc;
 using eShopLegacyMVC.Models;
 using eShopLegacyMVC.Services;
-using log4net;
+using Microsoft.Extensions.Logging;
 
 namespace eShopLegacyMVC.Controllers
 {
     public class CatalogController : Controller
     {
-        private static readonly ILog _log = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
-
+        private readonly ILogger<CatalogController> _log;
         private ICatalogService service;
 
-        public CatalogController(ICatalogService service)
+        public CatalogController(ICatalogService service, ILogger<CatalogController> logger)
         {
             this.service = service;
+            this._log = logger;
         }
 
         // GET /[?pageSize=3&pageIndex=10]
-        public ActionResult Index(int pageSize = 10, int pageIndex = 0)
+        public IActionResult Index(int pageSize = 10, int pageIndex = 0)
         {
-            _log.Info($"Now loading... /Catalog/Index?pageSize={pageSize}&pageIndex={pageIndex}");
+            _log.LogInformation($"Now loading... /Catalog/Index?pageSize={pageSize}&pageIndex={pageIndex}");
             var paginatedItems = service.GetCatalogItemsPaginated(pageSize, pageIndex);
             ChangeUriPlaceholder(paginatedItems.Data);
             return View(paginatedItems);
         }
 
         // GET: Catalog/Details/5
-        public ActionResult Details(int? id)
+        public IActionResult Details(int? id)
         {
-            _log.Info($"Now loading... /Catalog/Details?id={id}");
+            _log.LogInformation($"Now loading... /Catalog/Details?id={id}");
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return BadRequest();
             }
-            CatalogItem catalogItem = service.FindCatalogItem(id.Value);
+            CatalogItem? catalogItem = service.FindCatalogItem(id.Value);
             if (catalogItem == null)
             {
-                return HttpNotFound();
+                return NotFound();
             }
             AddUriPlaceHolder(catalogItem);
 
@@ -46,82 +45,78 @@ namespace eShopLegacyMVC.Controllers
         }
 
         // GET: Catalog/Create
-        public ActionResult Create()
+        public IActionResult Create()
         {
-            _log.Info($"Now loading... /Catalog/Create");
-            ViewBag.CatalogBrandId = new SelectList(service.GetCatalogBrands(), "Id", "Brand");
-            ViewBag.CatalogTypeId = new SelectList(service.GetCatalogTypes(), "Id", "Type");
+            _log.LogInformation($"Now loading... /Catalog/Create");
+            ViewBag.CatalogBrandId = new Microsoft.AspNetCore.Mvc.Rendering.SelectList(service.GetCatalogBrands(), "Id", "Brand");
+            ViewBag.CatalogTypeId = new Microsoft.AspNetCore.Mvc.Rendering.SelectList(service.GetCatalogTypes(), "Id", "Type");
             return View(new CatalogItem());
         }
 
         // POST: Catalog/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "Id,Name,Description,Price,PictureFileName,CatalogTypeId,CatalogBrandId,AvailableStock,RestockThreshold,MaxStockThreshold,OnReorder")] CatalogItem catalogItem)
+        public IActionResult Create([Bind("Id,Name,Description,Price,PictureFileName,CatalogTypeId,CatalogBrandId,AvailableStock,RestockThreshold,MaxStockThreshold,OnReorder")] CatalogItem catalogItem)
         {
-            _log.Info($"Now processing... /Catalog/Create?catalogItemName={catalogItem.Name}");
+            _log.LogInformation($"Now processing... /Catalog/Create?catalogItemName={catalogItem.Name}");
             if (ModelState.IsValid)
             {
                 service.CreateCatalogItem(catalogItem);
                 return RedirectToAction("Index");
             }
 
-            ViewBag.CatalogBrandId = new SelectList(service.GetCatalogBrands(), "Id", "Brand", catalogItem.CatalogBrandId);
-            ViewBag.CatalogTypeId = new SelectList(service.GetCatalogTypes(), "Id", "Type", catalogItem.CatalogTypeId);
+            ViewBag.CatalogBrandId = new Microsoft.AspNetCore.Mvc.Rendering.SelectList(service.GetCatalogBrands(), "Id", "Brand", catalogItem.CatalogBrandId);
+            ViewBag.CatalogTypeId = new Microsoft.AspNetCore.Mvc.Rendering.SelectList(service.GetCatalogTypes(), "Id", "Type", catalogItem.CatalogTypeId);
             return View(catalogItem);
         }
 
         // GET: Catalog/Edit/5
-        public ActionResult Edit(int? id)
+        public IActionResult Edit(int? id)
         {
-            _log.Info($"Now loading... /Catalog/Edit?id={id}");
+            _log.LogInformation($"Now loading... /Catalog/Edit?id={id}");
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return BadRequest();
             }
-            CatalogItem catalogItem = service.FindCatalogItem(id.Value);
+            CatalogItem? catalogItem = service.FindCatalogItem(id.Value);
             if (catalogItem == null)
             {
-                return HttpNotFound();
+                return NotFound();
             }
             AddUriPlaceHolder(catalogItem);
-            ViewBag.CatalogBrandId = new SelectList(service.GetCatalogBrands(), "Id", "Brand", catalogItem.CatalogBrandId);
-            ViewBag.CatalogTypeId = new SelectList(service.GetCatalogTypes(), "Id", "Type", catalogItem.CatalogTypeId);
+            ViewBag.CatalogBrandId = new Microsoft.AspNetCore.Mvc.Rendering.SelectList(service.GetCatalogBrands(), "Id", "Brand", catalogItem.CatalogBrandId);
+            ViewBag.CatalogTypeId = new Microsoft.AspNetCore.Mvc.Rendering.SelectList(service.GetCatalogTypes(), "Id", "Type", catalogItem.CatalogTypeId);
             return View(catalogItem);
         }
 
         // POST: Catalog/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Exclude = "PictureUri,CatalogType,CatalogBrand")] CatalogItem catalogItem)
+        public IActionResult Edit([Bind(include: "Id,Name,Description,Price,PictureFileName,CatalogTypeId,CatalogBrandId,AvailableStock,RestockThreshold,MaxStockThreshold,OnReorder")] CatalogItem catalogItem)
         {
-            _log.Info($"Now processing... /Catalog/Edit?id={catalogItem.Id}");
+            _log.LogInformation($"Now processing... /Catalog/Edit?id={catalogItem.Id}");
             if (ModelState.IsValid)
             {
                 service.UpdateCatalogItem(catalogItem);
                 return RedirectToAction("Index");
             }
-            ViewBag.CatalogBrandId = new SelectList(service.GetCatalogBrands(), "Id", "Brand", catalogItem.CatalogBrandId);
-            ViewBag.CatalogTypeId = new SelectList(service.GetCatalogTypes(), "Id", "Type", catalogItem.CatalogTypeId);
+            ViewBag.CatalogBrandId = new Microsoft.AspNetCore.Mvc.Rendering.SelectList(service.GetCatalogBrands(), "Id", "Brand", catalogItem.CatalogBrandId);
+            ViewBag.CatalogTypeId = new Microsoft.AspNetCore.Mvc.Rendering.SelectList(service.GetCatalogTypes(), "Id", "Type", catalogItem.CatalogTypeId);
             return View(catalogItem);
         }
 
         // GET: Catalog/Delete/5
-        public ActionResult Delete(int? id)
+        public IActionResult Delete(int? id)
         {
-            _log.Info($"Now loading... /Catalog/Delete?id={id}");
+            _log.LogInformation($"Now loading... /Catalog/Delete?id={id}");
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return BadRequest();
             }
-            CatalogItem catalogItem = service.FindCatalogItem(id.Value);
+            CatalogItem? catalogItem = service.FindCatalogItem(id.Value);
             if (catalogItem == null)
             {
-                return HttpNotFound();
+                return NotFound();
             }
             AddUriPlaceHolder(catalogItem);
 
@@ -131,17 +126,20 @@ namespace eShopLegacyMVC.Controllers
         // POST: Catalog/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public IActionResult DeleteConfirmed(int id)
         {
-            _log.Info($"Now processing... /Catalog/DeleteConfirmed?id={id}");
-            CatalogItem catalogItem = service.FindCatalogItem(id);
-            service.RemoveCatalogItem(catalogItem);
+            _log.LogInformation($"Now processing... /Catalog/DeleteConfirmed?id={id}");
+            CatalogItem? catalogItem = service.FindCatalogItem(id);
+            if (catalogItem != null)
+            {
+                service.RemoveCatalogItem(catalogItem);
+            }
             return RedirectToAction("Index");
         }
 
         protected override void Dispose(bool disposing)
         {
-            _log.Debug($"Now disposing");
+            _log.LogDebug($"Now disposing");
             if (disposing)
             {
                 service.Dispose();
@@ -159,7 +157,7 @@ namespace eShopLegacyMVC.Controllers
 
         private void AddUriPlaceHolder(CatalogItem item)
         {
-            item.PictureUri = this.Url.RouteUrl(PicController.GetPicRouteName, new { catalogItemId = item.Id }, this.Request.Url.Scheme);            
+            item.PictureUri = Url.RouteUrl(PicController.GetPicRouteName, new { catalogItemId = item.Id }, Request.Scheme) ?? string.Empty;
         }
     }
 }
